Tidy handling of SortOption type and the mapping of Sort parameter to SortOption
